requirements.txt
django
djangorestframework
django-filter
django-allauth
django-rest-auth